requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-jose